+++ src/zeusdb_vector_database/__init__.py
@@ -1,7 +1,7 @@
 """
 ZeusDB Vector Database Module
 """
-__version__ = "0.1.1"
+__version__ = "0.1.2"
 
 from .vector_database import VectorDatabase # imports the VectorDatabase class from the vector_database.py file